+++ deletetrip.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 include 'connection.php';
  error_reporting(0);
   $id = $_POST['id'];
@@ -50,28 +51,37 @@ if (mysqli_query($conn, $sql)) {
 		</div>
 		<div class="nav">
 			<ol>
-			<li><a href="index.php">Strona główna</a>
-			</li>
-			<li><a href="#">Przeglądaj oferty</a>
-			</li>
-			<?php
-			if(!isset($_SESSION['isAdmin']))
-				echo "<li><a href='login.php'>Zaloguj się</a></li>";
-			?>
-			<li><a href="signup.php">Zarejestruj się</a>
-			</li>
-			<li><a href="addtrip.php">Dodaj wycieczkę</a>
-			</li>
-			<?php
-			if($_SESSION['isAdmin'] == 1)
-				echo "<li><a href='deletetrip.php'>Usuń wycieczkę</a>"
-			?>
-			</li>
-			<?php
-			if(isset($_SESSION['isAdmin']))
-				echo "<li><a href='logout.php'>Wylogj</a></li>";
-			?>
-			
+				<li><a href="index.php">Strona główna</a>
+				<?php
+				if(!isset($_SESSION['isAdmin']))
+					echo "<li><a href='login.php'>Zaloguj się</a>";
+				?>
+				<?php
+				if(!isset($_SESSION['isAdmin']))
+				echo "<li><a href='signup.php'>Zarejestruj się</a>"
+				?>
+				<?php
+				if(isset($_SESSION['isAdmin']))
+				echo "<li><a href='addtrip.php'>Dodaj wycieczkę</a>"
+				
+				?>
+				<?php
+				if(isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] == 1)
+					echo "<li><a href='deletetrip.php'>Usuń wycieczkę</a>"
+				?>
+				<?php
+				if(isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] == 1)
+					echo "<li><a href='deleteuser.php'>Usuń użytkownika</a>"
+				?>
+				<?php
+				if(isset($_SESSION['isAdmin']) && $_SESSION['isAdmin'] == 1)
+					echo "<li><a href='edittrip.php'>Edytuj wycieczkę</a>"
+				?>
+				<?php
+				if(isset($_SESSION['isAdmin']))
+					echo "<li><a href='logout.php'>Wyloguj</a>";
+				?>
+				
 			</ol>
 		
 		</div>
@@ -81,6 +91,21 @@ if (mysqli_query($conn, $sql)) {
                 Podaj id wycieczki: <input type="text" name="id" required><br>			
                 <input type="submit" value="Usuń wycieczkę" name="deletetrip">
             </form>	
+			
+			<?php
+			include 'connection.php';
+			$sql = "SELECT * FROM trip";
+			$result = mysqli_query($conn, $sql);
+
+			if (mysqli_num_rows($result) > 0) {
+		   
+			while($row = mysqli_fetch_assoc($result)) {
+			echo  " <br> " . "ID: " . $row["id"]. "<br>" . "  Cel: " . $row["cel"].  " <br> " .  "Opis: " . $row["opis"] .  "<br>" . "Czar trwania: " . $row["trwanie"]. " dni"."<br><br>";
+			 }
+			} else {
+			echo "<h3><center>No user data found!<center></h3>";
+			}
+			?>
 		</div>
